+++ src/modules/gameAPI.js
@@ -1,0 +1,4 @@
+const gameId = 'RQlljNVPCPOWeSNibCxx';
+const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/';
+
+export { gameId, url };
